// src/enums/resource-type.ts
export enum ResourceType {
    Environment = 'environment',
    Router = 'router',
    Add = 'add',
    List = 'list',
    Panel = 'panel',
    PanelForm = 'panel form',
    Profile = 'profile',
    Query = 'query',
    QueryRoot = 'query root',
    Services = 'services',
    RootModule = 'root module',

}